+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 6.2.0 (Apr 15, 2022). Tested on Artifactory 7.35.2
+
+BUG FIXES:
+
+* resource/artifactory_pull_replication: Make `password` attribute configurable. `url`, `username`, and `password` attributes must be set together when use with remote repository. [GH-411]
+* resource/artifactory_push_replication: Make `password` attribute configurable. `url`, `username`, and `password` attributes are now required to match Artifactory API requirements [GH-411]
+
 ## 6.1.3 (Apr 12, 2022)
 
 BUG FIXES: